src/lib_project/pyproject.toml
[tool.poetry]
name = "lib_project"
version = "0.1.0"
description = "Library for common project management code"
authors = ["Till Speicher <dev@tillspeicher.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
hydra-core = "^1.3.2"
plotly = "^5.17.0"
paramiko = "^3.3.1"
nbconvert = "^7.9.2"
ipython = "^8.16.1"
nbformat = "^5.9.2"
jupyterlab = "^4.0.7"
ipywidgets = "^8.1.1"
pyarrow = "^16.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pre-commit = "^3.4.0"

[tool.poetry.group.edit]
optional = true

[tool.poetry.group.edit.dependencies]
neovim = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | \.mypy_cache
    | build
    | dist
)/
'''

[tool.isort]
atomic = true
profile = "black"
line_length = 80
skip_gitignore = true
lines_after_imports = 2
known_first_party = ["lib_project", "lib_dl_base"]
skip = [".gitignore"]

[tool.flake8]
max-line-length = 81
max-doc-length = 81
max-complexity = 10
ignore = ["E203", "E266", "E501", "W503", "F403", "F401", "E402", "C901"]
per-file-ignores = ["__init__.py:F401", "tests/*:S101"]
exclude = [".git", "__pycache__", "dist"]

[tool.pyright]
include = ["lib_project"]
executionEnvironments = [{ root = "tests", extraPaths = ["."] }]
reportMissingImports = "true"
pythonVersion = "3.10"
# The setup.sh script creates a symbolic link called .venv pointing to
# the virtual environment used by poetry
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = ["."]
testpaths = ["tests"]
